+++ src/cars/Saab95.java
@@ -17,7 +17,7 @@ public class Saab95 extends Car{
     /**
      * Instantiates a new <em>red</em> Saab95 with <em>two</em> doors, an engine power of <em>125</em>, and <em>disabled</em> turbo
      */
-    public Saab95(int x, int y){
+    public Saab95(double x, double y){
         super("Saab95", Color.red, 125, 2, Math.PI / 8, 40, 40, x, y);
         turboOn = false;
     }
