+++ src/cars/Scania.java
@@ -29,7 +29,7 @@ public class Scania extends Car {
     /**
      * Instantiates a new Scania.
      */
-    public Scania(int x, int y){
+    public Scania(double x, double y){
         super("Scania", Color.red, 250, 2, Math.PI / 8, 100, 300, x, y);
     }
     public Scania(){
