+++ src/cars/Volvo240.java
@@ -15,7 +15,7 @@ public class Volvo240 extends  Car{
     /**
      * Instantiates a new <em>black</em> Volvo240, with <em>four</em> doors and an engine power of 100
      */
-    public Volvo240(int x, int y){
+    public Volvo240(double x, double y){
         super("Volvo240", Color.black, 100, 4, Math.PI / 10, 40, 40, x, y);
     }
     public Volvo240(){
